wip feat: support buffer-write as part of buffer-create
I've long sought a nice convenience for this. I think I have it
now. buffer-create needs a context and a size in bytes. buffer-write
needs a srfi-4 vector and a command-queue.

But it's possible to get a context given a command-queue, and we can
leverage that here. So, supplying buffer-create with a command-queue
instead of context gives a couple of really nice possibilities:

1. we have all we need: context, command-queue, typt tag and size in
   bytes

2. we can issue buffer-write inside buffer-create

3. this should hopefully give the user an intuitive API, where
   supplying a srfi-4 vector to buffer-create will create a buffer
   with the content of that srfi-4 vector. the old API only used the
   srfi-4 for its type tag which was probably very confusing.

4. buffer-write can be used as before: given a context, you need to
   supply an size in number of bytes just as before.

Still a small TODO when buffer-create is supplied a buffer object (we
should copy its type tag and content).