bugfix: platform-devices not always returning all devices
aouch! this has been lying around for a long time. I noticed something
was wrong when clinfo -l would give me results, but platform-devices
wouldn't. I noticed it worked when I specified 'gpu instead of the
default 'all. So I checked how clinfo did it - and it was doing
exactly what I was doing.

Then I saw, in cl.h:

     ...
    typedef cl_ulong            cl_bitfield;
     ...
    typedef cl_bitfield         cl_device_type;
     ...

I must have missed that. Since

    #define CL_DEVICE_TYPE_ALL                          0xFFFFFFFF

this will be truncated when turned to a signed integer 32.